requirements.txt
streamlit==1.36.0
streamlit-feedback==0.1.3
watchdog==4.0.1
protobuf==3.20.*
requests
pillow
python-dotenv==1.0.1
SQLAlchemy==2.0.25
psycopg2==2.9.9